requirements.txt
google-generativeai
python-dotenv
streamlit
pandas

faker==25.0
faiss-cpu>=1.8
llama-index>=0.11.0
llama-index-embeddings-google-genai>=0.1.7
llama-index-vector-stores-faiss>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
pytest>=7.0
sentence-transformers>=2.2.0
plotly>=5.17.0 